install:deps: Stop weird call-back-to-self in resolveWithNewModule
We were doing a weird thing where we used a package.json field _installable_
to check to see if we'd checked for platform compatibility, and if not did
so.  But this was the only place that was ever done so there was no reason to
implement it in such an obfuscated manner.

Instead it now just directly checks and then records that its done so on the
node object with `knownInstallable`.  This is useful to know because modules
expanded via shrinkwrap don't go through this– `inflateShrinkwrap` does not
currently have any rollback semantics and so checking this sort of thing there
is unhelpful.