install/actions: Check installability of modules from shrinkwrap
Modules that came into the tree via shrinkwrap won't have had this
determined in advance so we check here and take advantage of the
optional rollback semantics to skip them.

Plus if it was a regular dependency we can just fail out now instead of
trying and failing to install the dep.

Fixes: #13394